+++ etna/transforms/missing_values/resample.py
@@ -135,13 +135,13 @@ class ResampleWithDistributionTransform(PerSegmentWrapper):
         self.in_column = in_column
         self.distribution_column = distribution_column
         self.inplace = inplace
-        self.out_column = self._get_out_column(out_column)
+        self.out_column = out_column
         super().__init__(
             transform=_OneSegmentResampleWithDistributionTransform(
                 in_column=in_column,
                 distribution_column=distribution_column,
                 inplace=inplace,
-                out_column=self.out_column,
+                out_column=self._get_out_column(self.out_column),
             )
         )