+++ etna/transforms/outliers/base.py
@@ -9,6 +9,7 @@ import pandas as pd
 
 from etna.datasets import TSDataset
 from etna.transforms.base import Transform
+from etna.transforms.utils import check_new_segments
 
 
 class OutliersTransform(Transform, ABC):
@@ -26,6 +27,7 @@ class OutliersTransform(Transform, ABC):
         self.in_column = in_column
         self.outliers_timestamps: Optional[Dict[str, List[pd.Timestamp]]] = None
         self.original_values: Optional[Dict[str, List[pd.Timestamp]]] = None
+        self._fit_segments: Optional[List[str]] = None
 
     def _save_original_values(self, ts: TSDataset):
         """
@@ -61,6 +63,7 @@ class OutliersTransform(Transform, ABC):
         ts = TSDataset(df, freq=pd.infer_freq(df.index))
         self.outliers_timestamps = self.detect_outliers(ts)
         self._save_original_values(ts)
+        self._fit_segments = ts.segments
 
         return self
 
@@ -75,14 +78,25 @@ class OutliersTransform(Transform, ABC):
 
         Returns
         -------
-        result: pd.DataFrame
+        result:
             dataframe with in_column series with filled with NaNs
+
+        Raises
+        ------
+        ValueError:
+            If transform isn't fitted.
+        NotImplementedError:
+            If there are segments that weren't present during training.
         """
         if self.outliers_timestamps is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling transform method.")
         result_df = df.copy()
-        for segment in df.columns.get_level_values("segment").unique():
-            result_df.loc[self.outliers_timestamps[segment], pd.IndexSlice[segment, self.in_column]] = np.NaN
+        segments = df.columns.get_level_values("segment").unique().tolist()
+        check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
+        for segment in segments:
+            # to locate only present indices
+            segment_outliers_timestamps = result_df.index.intersection(self.outliers_timestamps[segment])
+            result_df.loc[segment_outliers_timestamps, pd.IndexSlice[segment, self.in_column]] = np.NaN
         return result_df
 
     def inverse_transform(self, df: pd.DataFrame) -> pd.DataFrame:
@@ -96,13 +110,22 @@ class OutliersTransform(Transform, ABC):
 
         Returns
         -------
-        result: pd.DataFrame
+        result:
             data with reconstructed values
+
+        Raises
+        ------
+        ValueError:
+            If transform isn't fitted.
+        NotImplementedError:
+            If there are segments that weren't present during training.
         """
         if self.original_values is None or self.outliers_timestamps is None:
             raise ValueError("Transform is not fitted! Fit the Transform before calling inverse_transform method.")
         result = df.copy()
-        for segment in self.original_values.keys():
+        segments = df.columns.get_level_values("segment").unique().tolist()
+        check_new_segments(transform_segments=segments, fit_segments=self._fit_segments)
+        for segment in segments:
             segment_ts = result[segment, self.in_column]
             segment_ts[segment_ts.index.isin(self.outliers_timestamps[segment])] = self.original_values[segment]
         return result